dlq/requirements.txt
google-cloud-pubsub==2.11.0
cassandra-driver==3.25.0
